--- tensorly/contrib/optimization/optim_candecomp_parafac.py
@@ -1,161 +1,0 @@
-import numpy as np
-import warnings
-
-import tensorly as tl
-from tensorly.random import check_random_state
-from tensorly.base import unfold
-from tensorly.kruskal_tensor import kruskal_to_tensor
-from tensorly.tenalg import khatri_rao
-from tensorly.contrib.optimization.optim_tensor_ls import least_squares_nway
-from tensorly.contrib.optimization.optim_tensor_ls import nn_least_squares_nway
-
-# Author: Jean Kossaifi <jean.kossaifi+tensors@gmail.com>
-# Author: Chris Swierczewski <csw@amazon.com>
-# Author: Sam Schneider <samjohnschneider@gmail.com>
-
-# Optimization version : Jeremy Cohen (alpha version for demo)
-
-
-
-def initialize_factors(tensor, rank, init='svd', svd='numpy_svd', random_state=None, non_negative=False):
-    """Initialize factors used in `parafac`.
-
-    The type of initialization is set using `init`. If `init == 'random'` then
-    initialize factor matrices using `random_state`. If `init == 'svd'` then
-    initialize the `m`th factor matrix using the `rank` left singular vectors
-    of the `m`th unfolding of the input tensor.
-
-    Parameters
-    ----------
-    tensor : ndarray
-    rank : int
-    init : {'svd', 'random'}, optional
-    svd : str, default is 'numpy_svd'
-        function to use to compute the SVD, acceptable values in tensorly.SVD_FUNS
-    non_negative : bool, default is False
-        if True, non-negative factors are returned
-
-    Returns
-    -------
-    factors : ndarray list
-        List of initialized factors of the CP decomposition where element `i`
-        is of shape (tensor.shape[i], rank)
-
-    """
-    rng = check_random_state(random_state)
-
-    if init == 'random':
-        factors = [tl.tensor(rng.random_sample((tensor.shape[i], rank)), **tl.context(tensor)) for i in range(tl.ndim(tensor))]
-        if non_negative:
-            return [tl.abs(f) for f in factors]
-        else:
-            return factors
-
-    elif init == 'svd':
-        try:
-            svd_fun = tl.SVD_FUNS[svd]
-        except KeyError:
-            message = 'Got svd={}. However, for the current backend ({}), the possible choices are {}'.format(
-                    svd, tl.get_backend(), tl.SVD_FUNS)
-            raise ValueError(message)
-
-        factors = []
-        for mode in range(tl.ndim(tensor)):
-            U, _, _ = svd_fun(unfold(tensor, mode), n_eigenvecs=rank)
-
-            if tensor.shape[mode] < rank:
-                # TODO: this is a hack but it seems to do the job for now
-                # factor = tl.tensor(np.zeros((U.shape[0], rank)), **tl.context(tensor))
-                # factor[:, tensor.shape[mode]:] = tl.tensor(rng.random_sample((U.shape[0], rank - tl.shape(tensor)[mode])), **tl.context(tensor))
-                # factor[:, :tensor.shape[mode]] = U
-                random_part = tl.tensor(rng.random_sample((U.shape[0], rank - tl.shape(tensor)[mode])), **tl.context(tensor))
-                U = tl.concatenate([U, random_part], axis=1)
-            if non_negative:
-                factors.append(tl.abs(U[:, :rank]))
-            else:
-                factors.append(U[:, :rank])
-        return factors
-
-    raise ValueError('Initialization method "{}" not recognized'.format(init))
-
-
-def parafac(tensor, rank, n_iter_max=100, init='svd', svd='numpy_svd',
-        tol=1e-8, method='ALS',
-            orthogonalise=False, random_state=None, verbose=False, return_errors=False):
-    """CANDECOMP/PARAFAC decomposition via alternating least squares (ALS)
-
-    Computes a rank-`rank` decomposition of `tensor` [1]_ such that,
-
-        ``tensor = [| factors[0], ..., factors[-1] |]``.
-
-    Parameters
-    ----------
-    tensor : ndarray
-    rank  : int
-        Number of components.
-    n_iter_max : int
-        Maximum number of iteration
-    init : {'svd', 'random'}, optional
-        Type of factor matrix initialization. See `initialize_factors`.
-    svd : str, default is 'numpy_svd'
-        function to use to compute the SVD, acceptable values in tensorly.SVD_FUNS
-    tol : float, optional
-        (Default: 1e-6) Relative reconstruction error tolerance. The
-        algorithm is considered to have found the global minimum when the
-        reconstruction error is less than `tol`.
-    random_state : {None, int, np.random.RandomState}
-    verbose : int, optional
-        Level of verbosity
-    return_errors : bool, optional
-        Activate return of iteration errors
-
-
-    Returns
-    -------
-    factors : ndarray list
-        List of factors of the CP decomposition element `i` is of shape
-        (tensor.shape[i], rank)
-    errors : list
-        A list of reconstruction errors at each iteration of the algorithms.
-
-    References
-    ----------
-    .. [1] tl.G.Kolda and B.W.Bader, "Tensor Decompositions and Applications",
-       SIAM REVIEW, vol. 51, n. 3, pp. 455-500, 2009.
-    """
-    if orthogonalise and not isinstance(orthogonalise, int):
-        orthogonalise = n_iter_max
-
-    factors = initialize_factors(tensor, rank, init=init, svd=svd, random_state=random_state)
-    rec_errors = []
-    norm_tensor = tl.norm(tensor, 2)
-
-    for iteration in range(n_iter_max):
-        if orthogonalise and iteration <= orthogonalise:
-            factor = [tl.qr(factor)[0] for factor in factors]
-
-     # TODO: precompute unfoldings
-        if method == 'ALS':
-            factors, rec_error = least_squares_nway(tensor, factors, rank, norm_tensor)
-        if method == 'NALS':
-            factors, rec_error = nn_least_squares_nway(tensor, factors, rank, norm_tensor)
-
-        if tol:
-            rec_errors.append(rec_error)
-
-            if iteration > 1:
-                if verbose:
-                    print('reconstruction error={}, variation={}.'.format(
-                        rec_errors[-1], rec_errors[-2] - rec_errors[-1]))
-
-                if tol and abs(rec_errors[-2] - rec_errors[-1]) < tol:
-                    if verbose:
-                        print('converged in {} iterations.'.format(iteration))
-                    break
-                    
-    if return_errors:
-        return factors, rec_errors
-    else:
-        return factors
-
-
